Expand environment variables for paths...
in configuration files. Currently there is no way to put relative
paths in the configuration file to specify workspace directory
and so on. To fix that expand the paths so environment variables
can be used, e.g. ${PWD}, ${HOME} etc.
It is also possible to use ${HOME} relative paths, e.g.
"~/build".

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>