Make it possible to not remove SSTATE_DIR at any circumstances
Add new "--retain-sstate" command line argument which guarantees
that SSTATE directory is not removed even when setting up a new build.

Normally if a new build is started (either with --with-do-build or
without --continue-build) SSTATE_DIR is removed. If retain argument
is provided it is guaranteed that SSTATE directory is not removed then.

This behaviour is intended for local (non build server) builds where state
caches use is preferred over clean environment for every day development.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>
Reviewed-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>