Catch exceptions generated inside script
It's better to hide exception traces from user, to reduce amount of
not relevant information on screen.
In case of error we see colored messages from yocto.
Also for regular usage it's enough to see just text of exception,
in case of debug it's always possible to get any possible traces.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>