Rework of process_pulls()
Calls to PyGitHub results in usage of GitHub API usage.
This API has limitations that may affect build in some networks.
We use PyGitHub in function that is not critical for build and
provides some better user experience. So it's OK to remove calls
to PyGitHub to avoid interruptions during build.

This patch removes function that gets list of available PR's from github.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>