Allow to use empty strings in local_conf section
This commit fixes incorrect handling of empty strings assigned
to variables in `local_conf` section of configuration.

If empty string ("") is assigned or nothing is specified after `=`,
config parser provide tuple with only one field. Attempt to
access a second field of tuple results in error.

From now we handle this situation and write empty string for
variable with no value.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>